requirements.txt
# System
PyQt5
PyQt-Fluent-Widgets[full]
openai
anthropic
google-generativeai
tiktoken
opencc
openpyxl
lxml
beautifulsoup4

# Tools
rich
tqdm
httpx
json-repair
python-rapidjson